+++ packages/abjad-web/src/app/layout.tsx
@@ -1,4 +1,6 @@
+import Head from 'next/head'
 import { Analytics } from '@vercel/analytics/react'
+import { SpeedInsights } from '@vercel/speed-insights/next'
 import { Inter } from 'next/font/google'
 import type { Metadata } from 'next'
 import React from 'react'
@@ -12,9 +14,19 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
 	return (
 		<html lang="en">
+			<Head>
+				<link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png"/>
+				<link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png"/>
+				<link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png"/>
+				<link rel="manifest" href="/site.webmanifest"/>
+				<link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5"/>
+				<meta name="msapplication-TileColor" content="#da532c"/>
+				<meta name="theme-color" content="#ffffff"/>
+			</Head>
 			<body className={inter.className}>
 				{children}
 				<Analytics/>
+				<SpeedInsights/>
 			</body>
 		</html>
 	)